extract pageDetectors to separate file
